New propietary license
The license is propietary, not asquel. Asquel is free of charge and free to distribute. All what you have with other licenses like GNU GPL, plus some other articles that let you have more freedom to use and redistribute the software, if you want to create a commercial derivative of asquel, you can do it (with my written permission).